refactor(utils): isSameMonth naming and importing code at test file
