+++ src/plugins/withDateProps.ts
@@ -1,12 +1,12 @@
 import { getDate } from 'date-fns'
 
 import { DateCell } from '../models'
-import { isSameDate, isSameMonth } from '../utils'
+import { isSameDate, isSameYearAndMonth } from '../utils'
 
 export default function withDateProps(baseDate: Date, cursorDate: Date) {
   return function <T extends DateCell>(cell: T) {
     const { value: targetDate } = cell
-    const isCurrentMonth = isSameMonth(cursorDate, targetDate)
+    const isCurrentMonth = isSameYearAndMonth(cursorDate, targetDate)
     const isCurrentDate = isSameDate(baseDate, targetDate)
 
     return {
