+++ src/utils/index.ts
@@ -1,7 +1,7 @@
 export { default as arrayOf } from './arrayOf'
 export { default as generateID } from './generateID'
 export { default as isSameDate } from './isSameDate'
-export { default as isSameMonth } from './isSameMonth'
+export { default as isSameYearAndMonth } from './isSameYearAndMonth'
 export { default as parseDate } from './parseDate'
 export * from './pipe'
 export { default as withKey } from './withKey'
