+++ src/utils/isSameYearAndMonth.ts
@@ -1,5 +1,5 @@
 import { isSameMonth as _isSameMonth, isSameYear } from 'date-fns'
 
-export default function isSameMonth(baseDate: Date, targetDate: Date) {
+export default function isSameYearAndMonth(baseDate: Date, targetDate: Date) {
   return _isSameMonth(targetDate, baseDate) && isSameYear(targetDate, baseDate)
 }
